+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add "reject" decorators and traits
 * Add concrete noop decorators
 * Reorganize namespaces
+* Add extra-lazy decorators
 
 ## 1.0.0